requirements.txt
opencv-python
numpy
pandas
pyyaml
pydantic>=2
tqdm
scipy
# Detection / tracking (optional until you switch to YOLO)
ultralytics
supervision
# Review UI
streamlit
